Remove Windows support.
With the move to tokio 1.0, it looks likely that tokio-serial will drop
support for Windows until tokio adds a Windows equivalent of `AsyncFd`.
(See https://github.com/berkowski/tokio-serial/pull/33)

Let's stop running CI on Windows in response. Not removing cfg platform
gates out of the code for now in case we are able to restore Windows as
target later on.